requirements.txt
streamlit
plotly
pandas
numpy
scikit-learn
scipy
statsmodels
langchain
langchain-google-genai
python-dotenv
tqdm
